--- haystack/v2/pyproject.toml
@@ -1,93 +1,0 @@
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[project]
-name = "haystack"
-description = "End to end NLP applications at scale."
-readme = "README.md"
-requires-python = ">=3.8"
-license = "Apache-2.0"
-keywords = []
-authors = [
-  { name = "ZanSara", email = "sara.zanzottera@deepset.ai" },
-]
-classifiers = [
-  "Development Status :: 4 - Beta",
-  "License :: Freely Distributable",
-  "License :: OSI Approved :: Apache Software License",
-  "Programming Language :: Python",
-  "Programming Language :: Python :: 3.8",
-  "Programming Language :: Python :: 3.9",
-  "Programming Language :: Python :: 3.10",
-  "Programming Language :: Python :: 3.11",
-  "Programming Language :: Python :: Implementation :: CPython",
-  "Programming Language :: Python :: Implementation :: PyPy",
-]
-dynamic = ["version"]
-dependencies = [
-  "canals",
-]
-[project.optional-dependencies]
-rest-api = [
-  "uvicorn",
-  "fastapi",
-  "python-multipart"  # To allow file upload
-]
-dev = [
-  "hatch",
-  "pre-commit",
-  "mypy",
-  "pylint==2.15.10",
-  "black[jupyter]==22.6.0",
-  "pytest",
-  "coverage",
-]
-
-[project.urls]
-Documentation = "https://github.com/deepset-ai/haystack#readme"
-Issues = "https://github.com/deepset-ai/haystack/issues"
-Source = "https://github.com/deepset-ai/haystack"
-
-[tool.hatch.version]
-path = "VERSION.txt"
-
-[tool.hatch.envs.default]
-dependencies = [
-  "pytest",
-  "pytest-cov",
-]
-[tool.hatch.envs.default.scripts]
-cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=haystack --cov=tests {args}"
-no-cov = "cov --no-cov {args}"
-
-[[tool.hatch.envs.test.matrix]]
-python = ["38", "39", "310", "311"]
-
-[tool.coverage.run]
-branch = true
-parallel = true
-omit = [
-  "haystack/__about__.py",
-]
-
-[tool.coverage.report]
-exclude_lines = [
-  "no cov",
-  "if __name__ == .__main__.:",
-  "if TYPE_CHECKING:",
-]
-
-[tool.black]
-line-length = 120
-
-[tool.pylint.'MESSAGES CONTROL']
-max-line-length=120
-good-names="e"
-max-args=10
-disable = [
-  "line-too-long",
-  "missing-class-docstring",
-  "missing-module-docstring",
-  "consider-iterating-dictionary",  # horrible style imho :)
-]
