+++ haystack/v2/rest_api/test/test_files.py
@@ -1,5 +1,4 @@
 import os
-import json
 
 import pytest
